+++ src/Symbol.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of the webmozart/glob package.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webmozart\Glob;
+
+/**
+ * Contains symbol constants.
+ *
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+final class Symbol
+{
+    /**
+     * Represents a literal "\" in a regular expression.
+     */
+    const BACKSLASH = '\\\\';
+
+    /**
+     * Represents a literal "*" in a regular expression.
+     */
+    const STAR = '\\*';
+
+    /**
+     * Matches a literal "\" when running a regular expression against
+     * another regular expression.
+     */
+    const E_BACKSLASH = '\\\\\\\\';
+
+    /**
+     * Matches a literal "*" when running a regular expression against
+     * another regular expression.
+     */
+    const E_STAR = '\\\\\\*';
+
+    private function __construct()
+    {
+    }
+}